requirements.txt
streamlit as st
numpy as np
tensorflow as tf
from PIL import Image
pandas as pd
plotly.express as px